+++ test/com/pacman/renderer/DefaultRendererTest.java
@@ -2,11 +2,10 @@ package com.pacman.renderer;
 
 import static org.mockito.Mockito.*;
 
-import java.awt.Point;
-
 import org.junit.Test;
 import org.newdawn.slick.Graphics;
 
+import com.pacman.entity.Point;
 
 public class DefaultRendererTest {
 
@@ -15,33 +14,34 @@ public class DefaultRendererTest {
 		Graphics g = mock(Graphics.class);
 		Renderer renderer = new DefaultRenderer();
 		Renderable renderable = mock(Renderable.class);
-		Point point = new Point(10, 10);
-		
+		Point point = new Point(10f, 10f);
+
 		when(renderable.getPosition()).thenReturn(point);
-		
+
 		renderer.render(renderable, g);
-		
+
 		verify(g).pushTransform();
-		verify(g).translate((float) point.getX(), (float) point.getY());
+		verify(g).translate(point.getX(), point.getY());
 		verify(renderable).draw(g);
 		verify(g).popTransform();
 	}
-	
+
 	@Test
-	public void shouldTranslateObjectWithGivenSpeedBeforeRendering() throws Exception {
+	public void shouldTranslateObjectWithGivenSpeedBeforeRendering()
+			throws Exception {
 		Graphics g = mock(Graphics.class);
 		Renderer renderer = new DefaultRenderer();
 		Renderable renderable = mock(Renderable.class);
-		Point point = new Point(10, 10);
+		Point point = new Point(10f, 10f);
 
 		when(renderable.getPosition()).thenReturn(point);
-		
+
 		renderer.render(renderable, g);
-		
+
 		verify(g).pushTransform();
-		verify(g).translate((float) point.getX(), (float) point.getY());
+		verify(g).translate(point.getX(), point.getY());
 		verify(renderable).draw(g);
 		verify(g).popTransform();
 	}
-	
+
 }
